requirements.txt
anyio==4.4.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
python-telegram-bot==21.4
sniffio==1.3.1
